api/requirements.txt
appdirs==1.4.3
asyncpg==0.22.0
attrs==19.3.0
black==19.10b0
certifi==2020.12.5
Click==7.0
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
pathspec==0.7.0
pydantic==1.8.1
PyJWT==2.1.0
python-dotenv==0.17.0
regex==2020.2.20
SQLAlchemy==1.4.11
starlette==0.13.6
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
uvicorn==0.13.4
wincertstore==0.2
